+++ psched/src/sched_proc_test2.c
@@ -4,20 +4,20 @@
 #define SLEEPT 500*1000
 
 
-int init2();
-int run2();
+int16_t init2();
+int16_t run2();
 
 //int main(void){
 //  init();
 //  run();
 //}
 
-int init2(){
+int16_t init2(){
   printf("initialized module 2...\n");
   return SUCCESS;
 }
 
-int run2(){
+int16_t run2(){
   while(1){
     printf("ready to rumble with module 2...\n");
     usleep(SLEEPT);
